backend/requirements.txt
Django==4.2
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
Pillow
boto3
supabase
django-storages==1.14.6
PyJWT>=2.8.0
requests>=2.31.0
dj-database-url>=1.2.0
numpy>=1.26.0
celery>=5.3.0

# WSGI HTTP server for running Django in production
gunicorn==21.2.0
 
# Redis client for Celery/Kombu
redis>=4.6.0
cairosvg==2.0.0

# Machine learning/runtime dependencies are installed in the worker image only.
# See `backend/requirements_worker.txt` for the ML packages (torch, torchvision,
# segmentation-models-pytorch, timm) so the web backend image stays small.
